# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Directorio de Cartago API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface MessageResponse {
  message: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  visitor: "visitor",
  business_owner: "business_owner",
  admin: "admin",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  phone?: string | null;
  createdAt: string;
  hasActiveSubscription: boolean;
}

export type RegisterRequestRole =
  (typeof RegisterRequestRole)[keyof typeof RegisterRequestRole];

export const RegisterRequestRole = {
  visitor: "visitor",
  business_owner: "business_owner",
} as const;

export interface RegisterRequest {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  phone?: string;
  role?: RegisterRequestRole;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface ForgotPasswordRequest {
  email: string;
}

export interface ResetPasswordRequest {
  token: string;
  /** @minLength 6 */
  password: string;
}

export interface UpdateProfileRequest {
  name?: string;
  phone?: string;
}

export interface Category {
  id: number;
  name: string;
  icon?: string | null;
  description?: string | null;
  businessCount: number;
}

export interface CreateCategoryRequest {
  name: string;
  icon?: string;
  description?: string;
}

export interface BusinessImage {
  id: number;
  url: string;
  isPrimary: boolean;
}

export type BusinessStatus =
  (typeof BusinessStatus)[keyof typeof BusinessStatus];

export const BusinessStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface Business {
  id: number;
  name: string;
  description?: string | null;
  address: string;
  phone?: string | null;
  whatsapp?: string | null;
  instagram?: string | null;
  facebook?: string | null;
  website?: string | null;
  googleMapsUrl?: string | null;
  googleMapsEmbed?: string | null;
  schedule?: string | null;
  categoryId?: number | null;
  categoryName?: string | null;
  ownerId: number;
  ownerName?: string | null;
  status: BusinessStatus;
  images: BusinessImage[];
  averageRating?: number | null;
  reviewCount: number;
  createdAt: string;
}

export interface CreateBusinessRequest {
  name: string;
  description?: string;
  address: string;
  phone?: string;
  whatsapp?: string;
  instagram?: string;
  facebook?: string;
  website?: string;
  googleMapsUrl?: string;
  schedule?: string;
  categoryId?: number;
  images?: string[];
}

export interface BusinessListResponse {
  businesses: Business[];
  total: number;
  page: number;
  totalPages: number;
}

export interface RejectRequest {
  reason: string;
}

export interface SubscriptionPlan {
  id: number;
  name: string;
  description?: string | null;
  price: number;
  durationDays: number;
  isActive: boolean;
}

export interface CreateSubscriptionPlanRequest {
  name: string;
  description?: string;
  price: number;
  durationDays: number;
  isActive?: boolean;
}

export interface Subscription {
  id: number;
  userId: number;
  userName?: string | null;
  planId: number;
  planName?: string | null;
  startDate: string;
  endDate: string;
  isActive: boolean;
  paymentReference?: string | null;
}

export interface SubscribeRequest {
  planId: number;
  paymentReference?: string;
}

export interface UserListResponse {
  users: User[];
  total: number;
  page: number;
  totalPages: number;
}

export type AdminUpdateUserRequestRole =
  (typeof AdminUpdateUserRequestRole)[keyof typeof AdminUpdateUserRequestRole];

export const AdminUpdateUserRequestRole = {
  visitor: "visitor",
  business_owner: "business_owner",
  admin: "admin",
} as const;

export interface AdminUpdateUserRequest {
  name?: string;
  email?: string;
  role?: AdminUpdateUserRequestRole;
  phone?: string;
}

export interface AdminStats {
  totalUsers: number;
  totalBusinesses: number;
  pendingBusinesses: number;
  approvedBusinesses: number;
  activeSubscriptions: number;
  totalCategories: number;
}

export type GetBusinessesParams = {
  search?: string;
  categoryId?: number;
  status?: GetBusinessesStatus;
  page?: number;
  limit?: number;
};

export type GetBusinessesStatus =
  (typeof GetBusinessesStatus)[keyof typeof GetBusinessesStatus];

export const GetBusinessesStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export type AdminGetUsersParams = {
  page?: number;
  limit?: number;
  role?: string;
};