# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Directorio de Cartago API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "200":
          description: Reset email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /admin/users:
    get:
      operationId: adminGetUsers
      tags: [admin]
      summary: List all users (admin only)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: role
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /admin/users/{id}:
    put:
      operationId: adminUpdateUser
      tags: [admin]
      summary: Update user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: adminDeleteUser
      tags: [admin]
      summary: Delete user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /admin/subscription-plans:
    post:
      operationId: createSubscriptionPlan
      tags: [admin]
      summary: Create subscription plan (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriptionPlanRequest"
      responses:
        "201":
          description: Plan created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionPlan"
  /admin/subscription-plans/{id}:
    put:
      operationId: updateSubscriptionPlan
      tags: [admin]
      summary: Update subscription plan (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriptionPlanRequest"
      responses:
        "200":
          description: Plan updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionPlan"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [visitor, business_owner, admin]
        phone:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        hasActiveSubscription:
          type: boolean
      required: [id, name, email, role, createdAt, hasActiveSubscription]
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        phone:
          type: string
        role:
          type: string
          enum: [visitor, business_owner]
      required: [name, email, password]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
    ForgotPasswordRequest:
      type: object
      properties:
        email:
          type: string
          format: email
      required: [email]
    ResetPasswordRequest:
      type: object
      properties:
        token:
          type: string
        password:
          type: string
          minLength: 6
      required: [token, password]
...
    SubscriptionPlan:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        durationDays:
          type: integer
        isActive:
          type: boolean
      required: [id, name, price, durationDays, isActive]
    CreateSubscriptionPlanRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        durationDays:
          type: integer
        isActive:
          type: boolean
      required: [name, price, durationDays]
...
    UserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [users, total, page, totalPages]
    AdminUpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [visitor, business_owner, admin]
        phone:
          type: string
...